Created Floating Carousel
Improved the Eagles Nest page with swiperjs.com's floating carousel per the figma. Desktop is working, however, mobile needs some changes. The breakpoint feature of Swiper needs to be used in order to fix sizing for various devices.

Lots of styling still needs to be done for Eagles Nest, and the eagle information tied to the floating carousel needs to be added.